Remove _newTemp from TryStartInspiration
Fucking really, Tynan?  I think the advertisement that TryStartInspiration would be deprecated a year ago was backwards, and it was NewTemp that was going to be deprecated in favor of TryStartInspiration, obviously because "NewTemp" implies that it's temporary...

This is literally the only change in the code you have to make in order to get everything to compile correctly.